handle login/register in the render.js
